Normalize WinSDK access mask types before combining fsync flags
The exact-1b5dbc8 native Windows gate stopped during compilation because
WinSDK exposes GENERIC_READ as UInt32 and GENERIC_WRITE as Int32. Cast
each operand to DWORD before bitwise OR, matching the existing native
AdvisoryLock adapter. The resulting access mask is unchanged.

Native diagnostic: AtomicWrite.swift:217, binary operator | cannot be
applied to UInt32 and Int32. No tests ran in that attempt; the native
wrapper exited 1 with unchanged exact source. git diff --check passes.
The next exact-commit native gate verifies this Windows-only correction.

+++ Sources/OpenGrokFileUtils/AtomicWrite.swift
@@ -214,7 +214,7 @@ public enum AtomicFile: Sendable {
         let rawHandle = native.withCString(encodedAs: UTF16.self) {
             CreateFileW(
                 $0,
-                DWORD(GENERIC_READ | GENERIC_WRITE),
+                DWORD(GENERIC_READ) | DWORD(GENERIC_WRITE),
                 DWORD(FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE),
                 nil,
                 DWORD(OPEN_EXISTING),
